Add dsl namespace to entities
To better encapsulate methods exposed to the user via the DSL and
not bleed other methods into the user DSL, this commit adds DSL
classes and objects to all entities (books, sections, steps).

Tangibly, this modifies the way that extensions are implemented.
Now methods that are targeted at the user DSL for an entity should
prepend to that entity's DSL module as opposed to the module itself.
Additionally, scopes in the DSL should be evaluated in the context of
the corresponding entity's DSL object. Finally, the parent object can
be accessed from the DSL's context by referencing the `parent` object.