Fix paranoid jump bug
When in paranoid mode, skip and jump fail in certain cases. The problem
occurs because the metadata[:start_at] value is set, but this new value
is only inherited by children and siblings of the current step, but not
parents or other distantly related nodes. This commit changes the value
of metadata[:start_at], but leaves the reference intact. This way
parents of the current step and other siblings continue to reference the
same value.