Add repo metadata
Due to the way metadata is generated and passed to each node in a
runbook, new data added to metadata is no available to sibling nodes or
their descendents (only the nodes children get access to any changes to
metadata). In order to add a way to pass data between nodes, this commit
adds a repo hash. Data added to the repo hash will be available to any
other node that is traversed when processing the book.