Fix bug to allow disabling paranoid mode during execution
Because each node gets a copy of its own metadata, modifying metadata
from one node will not propogate to the rest of the nodes. Hence,
modifying the paranoid metadata within one node does not globally
disable this value.

In order to fix this, I created a StickyHash which can have a Glue
object added to it. This glue object will persist its data and ensure it
is shared across all clones of the original StickyHash. Now metadata is
a StickyHash, and the paranoid key contains a Glue value.