Prevent loading repo file when start_at is passed nil
When initializing a runner, you may wish to pass nil to start_at for
example when pulling a value from a hash and not wanting to specify the
explicit default of "0". This commit converts a nil passed to runner to
the default value of "0". This has the effect of preventing an existing
repo file from being loaded when a nil start_at is passed and prevents the
runbook from crashing when running runbook with Ruby <= 2.2.2 and
passing nil.