Clear open pagers before sending tmux commands to panes
When sending commands to a tmux pane, there is no assertion that a
command prompt is ready on the pane to receive the sent command. Often
times a prompt is not available because a pager (more, less, etc.).
exits in that pane. This commit sends 'q C-u' before your command in an
effort to clear any open pager in the pane before your command is sent.

If a pager is open in the pane, this has the effect of sending 'q' to
the pager to exit it, and then it sends <ctrl>-u to the terminal which
has the effect of clearing all existing input (a noop). If no pager
exists, then a 'q' is sent to the prompt and <ctrl>-u is sent to
immediately clear the q and any other typed characters in the prompt.

Future work could require that tmux commands send this explicitly (a
backwards incompatible change), by specifying an additional parameter
for pre-command characters or an escape sequence.