Fix setup_layout race condition
Previous to this change, If you invoke a runbook and then switch panes,
when looking for the _layout_file that stores the layout, Runbook will
use the pane_pid and pane_id of the active pane as opposed to the pane
where Runbook is invoked.

With this change Runbook will look for the layout file based on the pane
of where the runbook is invoked.