Extract Runbook helpers to util
This commit moves Runbook.{entities,statements} to a separate
util/runbook.rb file. Additionally this commit adds the Runbook.runs
method which returns all Run modules currently defined.

This is necessary because extensions may need to add functionality to
already defined run modules, and as such it is helpful to get a list of
the defined run modules.