Add shared_variables extension
Currently it is a clumsy process to pass values between different
statements and entities. This commit streamlines this by setting
instance variables on each entity and statement from the runbooks repo,
and syncing any instance variables set back to the repo after each
object finishes executing and before a new object starts executing.

One thing to be aware of is that instance variables do not get synced to
the repo until after the statement or entity finishes or a new entity
starts executing. For this reason, behavior can be a little weird if you
do not consistently access values either via the repo or via instance
variables.

Because ivars (and corresponding attr_accessors) share a namespace with
each object's dsl, there is a potential for dsl methods to be clobbered.
This is something you should keep in mind when naming variables. In the
future it may be helpful to detect this and print a warning if it
happens.