Allow frozen values to be passed as start_at
This commit refactors the way start_at is set to use the Glue helper.
This allows us to avoid a few other work-arounds such as editing the
value of start_at in place and duplicating it before it is passed in so
that frozen variable can be passed in. Using the glue helper (which was
designed specifically for this case) helps prevent the necessity for
these workarounds.