Fix remote sudo bug
Currently executing remote commands as another user fails with the
following message:

`sudo: no tty present and no askpass program specified`

This commit fixes this problem by executing sudo commands with pty
set when enable_sudo_prompt is set and a user is specified.

This could potentially cause issues with commands that used to run
without a pty but now run with a pty. They could attempt to accept
user input under this new configuration. One workaround is to use
raw input and not specify a user. Additionally, you can disable the
sudo prompt and set up sudoers to allow the command.