Prompt to resume from previous pose when stored
This commit adds functionality to resume from the last position when a
runbook is restarted. This is meant to protect from the issue where a
runbook is restarted after it ungracefully exits and the previous state
is blown away because start_at was not passed.

Some concerns with this design:

1. It must save a new file at the beginning of each object. This adds
some overhead and we are already saving the repo file after each object.
Ultimately the code was cleaner managing two separate files as opposed
to combining them into a single file.

2. I am concerned around adding an unexpected prompt to the workflow and
also that the correct behavior is executed when the auto and noop flags
are passed. I decided to start from the beginning in the case of auto
and noop, because this is the least surprising behavior when resuming a
failed runbook. You don't expect it to automatically resume from it's
previous position. In order to avoid this ambiguity or odd behavior, it
is best to pass the start_at flag if you want to resume at a failed
step.