Update start_at_is_substep? definition
This commit updates start_at_is_substep's definition.

First, it returns true if metadata[:position] is
empty because only books should have an empty
position, and all objects are substeps of the book.

Second, it returns false if the current object is not
an entity. Non-entities do not have children so
start_at can't be a substep of the current position.