Allow jumping to previous steps in paranoid mode
Currently, when entering a step to jump to, if you enter an already-past
step, then that jump will effectively be ignored. This commit allows you
to jump back to a previous step.

It should be noted that hook behavior can be confusing and hard to
follow when jumping steps. When you jump to a previous step, around
hooks and after hooks will be executed for the current step and its
parents. Hooks will be executed for the step you jump to, and only after
hooks will be executed for parents of the step you jump to.

The control flow is a little weird for traversing backwards.
Specifically, there is some redundancy in the added control flow. Part
of this is for optimization, and part of this is for handling weird edge
cases such as jumping to the same step.

Additionally, two metadata values are required to be stored in order to
properly handle the control flow and prevent child statements of a book
(description, layout, etc.) from being re-executed when jumping to an
older step.

Also discovered and fixed a broken test ("renders code blocks").

Also discovered runbook files are evaluated twice if they cannot be
"loaded". This can cause hooks to execute multiple times.