Use previous values as defaults for ask statements
This commit sets previous values as defaults for ask statements.
For example, if you need to go back to a step to update an input
value, its default will be the input value that you previously
set. This is helpful for when you may need to go to a previous
step, but your input values do not need to change. This will
save you from having to retype these values.

The implementation for this is pretty clean and straightforward. My one
concern with it is that it introduces a backwards incompatibility if
people are resuming runbooks using auto and expecting previous defaults
to be set, but this seems like a relatively unlikely scenario and in the
majority of cases, you would want values set which you previously set
for this run, when possible.