Fix Thor usage bug
Previously I was using a non-invocation error from Thor. I believe the
error I was using to signify an unknown runbook file is intended for
internal use. I updated this to be an invocation error, which is an
error that signifies Thor was invoked incorrectly (with an unknown file
in this specific case).