Fix config file load ordering to support extension
Currently global, project, and user config files are loaded when the
runbook gem is required. If runbook configuration is extended to allow
for additional configuration from an external gem, these config files
cannot contain configuration for the runbook extension. This is because
the extension must be required after runbook, but the configuration has
already run before the extension is allowed to add its configuration
parameters.

The purpose of this commit is to update the way configuration is loaded
to avoid this issue. Because we can no longer load runbook configuration
when the gem is required, we need to find another point at which to load
the configuration that occurs after extension gems are required but
before runbook runs. Because runbook does not have a single entry point
(runbooks can be invoked from the command line, calling runners or
viewers, etc.), it is difficult to pinpoint when configuration should be
loaded.

This commit identifies four entrypoints for executing runbook
code: Runbook.book, Runbook.section, Runbook.step, and
Runbook.configure. Runbook attempts to load config at these entry points
with memoization, so that if config has already been loaded, it does not
attempt to load the config again.

At these entry points, all extensions should be required and no other
runbook code should have run. We have access to the 4 points of
configuration that should now be available: global, project, user, and
cli.