Do not prompt for steps without titles in paranoid mode
When running in paranoid mode, it is jarring to stop at a step without
any description. It is unclear whether you should continue without a
description of what the runbook is going to do next.

Steps without titles are typically brief and what is being executed
should be self-explanatory and not risky. Consequently, we skipped
prompting for any step that does not have a title. If you wish to still
prompt at a step, but do not want to add a title, you can initialize the
step with an empty string and the output will be the same, but the
runbook will still prompt to continue.

It may make sense to make this behavior configurable based on
preference.