Change runbook registration process
This is a backwards incompatible change. This changes the implementation
of Runbook.books from a hash to an array. Code that references
Runbook.books should interact with it as if its an array instead of
expecting it to be a hash. Tangibly this means accessing the last
element of the array, as opposed to pulling a book out based on a key.

Additionally, registering books by adding books to Runbook.books is no
longer necessary. Books are now automatically registered when they are
defined.

A hash implementation, where you have to register the book under a
specific name, and then retrieve the runbook under the same name is
overly complex and doesn't provide any benefit over a simpler solution.
As it's currently implemented for the CLI, one runbook gets registered
under the name of the file, and this same runbook gets retrieved by file
name magic and gets executed.

A much simpler implementation would have a runbook assigned to a
constant, and then use that constant when executing the runbook. It also
feels weird to store the exact runbook to be executed at a magic
constant.

Using a single constant to store the runbook to be executed seems to be
too restrictive of an implementation. Multiple runbooks can be defined
per file, and it seems important to be able to treat these runbooks
equally and provide the ability to work with all of these runbooks.

This commit auto-registers runbooks when they are defined. It would be
nice to allow the user to choose which runbooks get registered, but
convention over configuration would dictate that runbooks should
register themselves and we execute the last one that's defined. If users
do not want a runbook to be registered for execution, then they should
prevent the runbook from being defined in the first place. This would be
more performant, and there is no purpose in defining a runbook that won't
be run.

The CLI executes the last runbook that is registered. This behavior
could be changed in the future to execute _all_ runbooks that get
registered.