Fix jump backwards dynamic statement bug
Runbook currently has a bug where if you jump backwards and rexecute a
ruby_command statement that defines additional statements, the old
statements will continue to be executed. This is because dynamic
statements get added to the execution tree, and re-executing the
ruby_command adds additional new statements.

This commit ensures dynamically defined statements (statements defined
within a ruby_command) are only executed once.

This commit also covers the case of entities being defined dynamically,
though this functionality is discouraged.

In an ideal world, dynamic statements would be completely overwritten
when their ruby_command is re-executed, however this implementation is
considerably simpler. The downside is that now if you re-execute the
ruby_command before executing all of its dynamic statements, the new
statements will be executed and then the un-executed old statements will
be executed.