Refactor config validation in cli
This commit refactors the cli's config file validation so it does not
need to be defined for each command. This is particularly helpful when
defining commands that have subcommands because they cannot inline
validating configuration.

The downsides of this pull request are overriding the initialize method
and a less descriptive unknown config file error message. The benefits
of this commit are new commands automatically inherit config file
validation, and that this allows config file validation to be applied to
subcommand commands.