Add runbook generate command
This commit adds the generate command to the runbook cli and a generator
generator for generating new generators. Some of the odd aspects of this
commit include:

1) Config must be loaded before the generator cli is executed so
user-defined generators can be found. Because of this execution order,
non-commandline configs will be loaded twice, once to get the
configuration for the generator commands and again for the regular
runbook lifecycle.

2) Due to thor weirdness class_options for generators must be defined at
the generate command level and at the individual generator level. This
is because help is generated using the options at the command level, but
the actual behavior is controlled by the definitions on the generators
that extend thor group.