Make runbook state files only user-readable
This commit ensures that runbook state files (which are stored in /tmp)
are only readable and writeable by the current user. This has been added
as a security measure as potentially secure data such as passwords and
tokens could be stored in these files, and it is safest to ensure they
are not readable by other users of the system.