Provide run as an argument to ruby_command
Giving the ruby_command access to the run enables it to reference any
methods defined on the run. ruby_commands used to have the ability to
reference these static methods before ruby_command was updated to
execute in the context of the parent's dsl.

When referencing the run, it may be necessary to be defensive by
checking if methods are defined before executing them. This is because
you technically don't know what run will be used to execute your
runbook.

The before(:each) change is due to a bug when the ruby_command's parent
step has no items (which should happen in practice).