Add airbrussh context
This commit adds support for airbrussh context so that elapsed time at
each step is printed along with the step's title when sshkit commands
are executed. Additionally, command numbering is slightly more intuitive
as incrementing numbers are scoped to a step. Repeated,
non-consecutive commands receive new numbers (instead of the number of
the first call of the command).

There is still a little bit of incongruence between airbrussh and
Runbook. In particular, subsequent steps with the same title (think two
blank steps in a row) will not re-print the step's title. Additionally,
the elapsed time counter starts with the first sshkit command, not with
the start of the runbook.